+++ Changelog.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## Latest
+## 1.9a
 * Improved KOTH scoreboard and added disappearing smoke workaround
 * Added steersman zig-zag enable/disable function
 * Added "Hold Position" / "Resume Sailing" options to manually controlled steersman groups
 * Added random trigger flag and simplifed comms scripting utilities for HeLMS
 * Added respawn in zone methods for HeLMS
+* Added limited methods to show/hide drawings dynamically
 ## 1.8a
 * Created common comms management logic in HeLMS
 * Steersman comms control option added